video: console: Setfont breaks first Xserver start
References: 302010
Patch-Mainline: No, <dmueller: reposted to linux-console, Oct 4, 2017>

The patch prevents setfont from accessing vga registers on the card when
the card is in graphics mode KD_GRAPHICS as we assume, that someone else
(ie. the Xserver) is in charge of the HW in which case accessing the vga
registers may (at best) have no effect (not even the desired one) or (at
worst) interfer with settings the graphics driver has made.

Signed-off-by: Hannes Reinecke <hare@suse.de>