i2c: i2c-designware-platdrv: Allow a dynamic adap. nr without...
...an ACPI fwnode

Before this commit the i2c-designware-platdrv assumes that if the pdev
has an apci-companion it should use a dynamic adapter-nr and otherwise
it will use pdev->id as adapter-nr.

On some devices e.g. the Apollo Lake using Acer TravelMate Spin B118,
some of the LPSS i2c-adapters are enumerated through PCI and do
not have an ACPI fwnode. These devices are handled as mfd devices so
they end up using the i2c-designware-platdrv driver.

This results in the i2c-adapter being registered with the
mfd generated pdev->id as adapter-nr, which conflicts with existing
adapters, triggering a WARN(id < 0, "couldn't get idr") in
i2c-core-base.c and causing the adapter registration to fail.

This commit adds support for setting a "linux,use-dynamic-adapter-nr"
device property on the device to make i2c-designware-platdrv
use dynamic adapter-nrs on devices without an ACPI fwnode,
together with changes to
drivers/mfd/intel-lpss-pci.c to set this, this fixes the WARN.

Before this commit the setting of the adapter.nr was
somewhat convoluted, in the acpi_companion case it was set from
dw_i2c_acpi_configure, in the non acpi_companion case it was set from
dw_i2c_set_fifo_size() based on tx_fifo_depth not being set yet. This
commit also cleans this up.

BugLink: https://bugzilla.redhat.com/show_bug.cgi?id=1687065
Signed-off-by: Hans de Goede <hdegoede@redhat.com>