mfd: intel-lpss: Add "linux,use-dynamic-adapter-nr" device-prop to...
...i2c-devices

The i2c-designware-platdrv assumes that if the pdev has
an apci-companion it should use a dynamic adapter-nr and otherwise it
will use pdev->id as adapter-nr.

On some devices e.g. the Apollo Lake using Acer TravelMate Spin B118,
some of the LPSS i2c-adapters are enumerated through PCI and do
not have an ACPI fwnode. These devices are handled as mfd devices so
they end up using the i2c-designware-platdrv driver.

This results in the i2c-adapter being registered with the
mfd generated pdev->id as adapter-nr, which conflicts with existing
adapters, triggering a WARN(id < 0, "couldn't get idr") in
i2c-core-base.c and causing the adapter registration to fail.

To fix this the i2c-designware-platdrv now also supports a
"linux,use-dynamic-adapter-nr" device-property, which tells it to
use a dynamic adapter nr even if there is no ACPI fwnode for the device.

This commit adds this device-property to the properties list for PCI
enumerated LPSS i2c-adapters.

BugLink: https://bugzilla.redhat.com/show_bug.cgi?id=1687065
Signed-off-by: Hans de Goede <hdegoede@redhat.com>