acpi: acpica: Calling ObjectType on buffer field reports type integer
It seems that buffer fields (created via CreateField) are incorrectly
reported as being of type integer (via ObjectType) when they are small
enough to allow for that.

As far as I know all kernel versions are affected by this, I have
specifically checked 5.0 (5.0.2), 4.20, 4.19 and 4.18.

In more detail: On the Microsoft Surface Book 2, Surface Pro (2017),
Surface Pro 6, Surface Laptop and Surface Laptop 2 devices there is
the following piece of AML code:

     Local0 = ^^_SAN.RQST (0x02, One, One, Zero, One)
     If ((ObjectType (Local0) != 0x03))
     {
         // Error path
     }
     Else
     {
         // Success path
     }

Which executes a command (RQST) and then checks the type of the
returned field to determine if that command was successful (i.e.
returned a buffer object) or failed (i.e. returned any other type,
including integer). The buffer field that is being returned by RQST
is rated as follows:

     CreateField (RQBF, 0x20, Local3, ARB)
     Local4 = ARB /* \_SB_._SAN.RQSX.ARB_ */
     ...
     Return (Local4)

If the returned buffer field is small enough (smaller than
acpi_gbl_integer_byte_width), the error-path will always be executed.

The full DSDT for the Surface Book 2 can be found here:
https://github.com/qzed/surfacebook2-dsdt/blob/fa0ca7c7cba8fb0da1e79c282f9a9f4a12eaaa17/dsdt.dsl#L15396

I have attached a patch (for 5.0) that fixes this issue and has been
in use on the affected devices for a few months now via

     https://github.com/qzed/linux-surfacegen5-acpi and
     https://github.com/jakeday/linux-surface/

I'm not aware of any issues regarding the patch, but then again this
has, to my knowledge, only been tested on Microsoft Surface devices.


Signed-off-by: Maximilian Luz <luzmaximilian@gmail.com>