ASoC: intel: cht_bsw_max98090_ti: Enable codec clock once and...
...keep it enabled

Users have been seeing sound stability issues with max98090 codecs
since:
commit 648e921888ad ("clk: x86: Stop marking clocks as CLK_IS_CRITICAL")

At first that commit broke sound for Chromebook Swanky and Clapper
models, the problem was that the machine-driver has been
controlling the wrong clock on those models since support for them
was added. This was hidden by clk-pmc-atom.c keeping the actual
clk on unconditionally.

With the machine-driver controlling the proper clock, sound works
again but we are seeing bug reports describing it as: low volume,
"sounds like played at 10x speed" and instable.

When these issues are hit the following message is seen in dmesg:
"max98090 i2c-193C9890:00: PLL unlocked".

Attempts have been made to fix this by inserting a delay between
enabling the clk and enabling and checking the pll, but this has
not helped.

It seems that if we ever disable the clk, the pll looses its lock
and then we get various issues.

This commit fixes this by enabling the clock once at probe time,
in essence restoring the old behavior of clk-pmc-atom.c always
keeping the clk on, but then only on machines with a max98090 codec.

Fixes: 648e921888ad ("clk: x86: Stop marking clocks as CLK_IS_CRITICAL")
Signed-off-by: Hans de Goede <hdegoede@redhat.com>