hmat: Register attributes for memory hot add
Some of the memory nodes described in HMAT may not be online at the
time the hmat subsystem parses their nodes' attributes. Should the node
be set to online later, as can happen when using PMEM as RAM after boot,
the nodes will be missing their initiator links and performance
attributes.

Regsiter a memory notifier callback and register the memory attributes
the first time its node is brought online if it wasn't registered.

Signed-off-by: Keith Busch <keith.busch@intel.com>