hid: remove NO_D3 flag for non Cherry Trail(CHT) [v2]
The code base NO_D3 flag is especially set for CHT, which is a tablet
platform, it have PMC changes and do not need to enter D3.
The newer platforms, which need ISH to enter D3, if have this NO_D3
flag set they can never enter D3, so remove it.

Signed-off-by: Song Hongyan <hongyan.song@intel.com>