mac80211: add support for the ADDBA extension element
HE allows peers to negotiate the aggregation fragmentation level to be
used during transmission. The level can be 1-3. The Ext element is added
behind the ADDBA request inside the action frame. The responder will
then reply with the same level or a lower one if the requested one is
not supported.
This patch only handles the negotiation part as the ADDBA frames get
passed to the ATH11k firmware, which does the rest of the magic for us
aswell as generating the requests.

Signed-off-by: Shashidhar Lakkavalli <slakkavalli@datto.com>
Signed-off-by: John Crispin <john@phrozen.org>